Enemy Orbit Works
I accidentally got rid of the movement script file and re-added it, and had to delete it from it being somewhere else in the files.
I also changed the multiplier for FacePlayer rotation to be twice as fast. (Enemies that are farther away aren't looking right at the player and are looking off to the side a little bit, I tried to make the rotation faster to counter-act that, but it still happens so I'm leaving it alone.)